NOVEMBER/PGP_v1/PGP_SERVER_v1/requirements.txt
# TelePay10-26 Dependencies

# Telegram Bot Framework
python-telegram-bot>=20.0

# Flask Web Framework
Flask>=3.0.0
gunicorn>=21.2.0

# Flask Security Extensions
flask-wtf>=1.2.0          # CSRF protection
flask-talisman>=1.1.0     # Security headers (CSP, HSTS, etc.)

# Database - SQLAlchemy ORM
sqlalchemy>=2.0.0

# Database - PostgreSQL Driver (pg8000 is pure Python, no C dependencies)
pg8000>=1.31.1
psycopg2-binary>=2.9.9

# Database - Cloud SQL Python Connector
cloud-sql-python-connector>=1.5.0

# HTTP Client
httpx>=0.25.0
requests>=2.31.0

# Async Support
asyncio-mqtt>=0.16.0
nest-asyncio>=1.5.0

# Security - Cryptography
cryptography>=41.0.0

# Environment Variables
python-dotenv>=1.0.0

# Google Cloud
google-cloud-secret-manager>=2.16.0
google-cloud-logging>=3.5.0

# Utilities
pytz>=2023.3
